requirements.txt
llama-cpp-python==0.3.4 --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cu121/
openai
fastapi
langchain
faiss-cpu
tabulate
pandas
numpy
pypdf2
openpyxl
uvicorn